# apps/server/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "AppController_getHello",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "tags": ["App"]
      }
    },
    "/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateUserDto" }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["Users"]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "tags": ["Users"]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Users"]
      },
      "patch": {
        "operationId": "UsersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateUserDto" }
            }
          }
        },
        "responses": { "200": { "description": "" } },
        "tags": ["Users"]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Users"]
      }
    },
    "/files/upload": {
      "post": {
        "operationId": "FilesController_uploadFile",
        "parameters": [],
        "responses": { "200": { "description": "" } },
        "tags": ["Files"]
      }
    }
  },
  "info": {
    "title": "Comcamp36 API",
    "description": "Our lovely reference for Comcamp36 Website",
    "version": "1.0",
    "contact": {}
  },
  "tags": [],
  "servers": [],
  "components": {
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "google_id": { "type": "string" },
          "email": { "type": "string" }
        },
        "required": ["google_id", "email"]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "fullname": { "type": "string" },
          "age": { "type": "string" },
          "birth": { "type": "string" },
          "gender": { "type": "string" },
          "religion": { "type": "string" },
          "blood_group": { "type": "string" },
          "graduation": { "type": "string" },
          "school": { "type": "string" },
          "course": { "type": "string" },
          "telephone": { "type": "string" },
          "medical_coverage": { "type": "string" },
          "chronic_diseas": { "type": "string" },
          "self_medicine": { "type": "string" },
          "drug_allergic": { "type": "string" },
          "food_allergic": { "type": "string" },
          "perfer_food": { "type": "string" },
          "address": { "type": "string" },
          "home_phone_tel": { "type": "string" },
          "comcamp_attendance": { "type": "string" },
          "shirt_size": { "type": "string" },
          "everyday_attendence": { "type": "string" },
          "has_laptop": { "type": "string" },
          "travel": { "type": "string" },
          "parent_fullname": { "type": "string" },
          "parent_relation": { "type": "string" },
          "parent_phone": { "type": "string" },
          "has_submit_answer": { "type": "string" }
        }
      }
    }
  }
}